# backend/docs/openapi/api.yaml
...
openapi: 3.0.3
info:
    description: |
        Machine-generated OpenAPI specification for AppOS APIs, with ext routes generated from route code using group-matrix as single source of truth.
    title: AppOS API
    version: 0.1.0
...
components:
    schemas:
        ErrorEnvelope:
            properties:
                code:
                    type: integer
                message:
                    type: string
            required:
                - message
            type: object
        GenericRequest:
            additionalProperties: true
            description: Generic request payload placeholder (refine per endpoint)
            type: object
...
        SuccessEnvelope:
            additionalProperties: true
            description: Generic success payload envelope
            example:
                ok: true
            type: object
    securitySchemes:
        bearerAuth:
            bearerFormat: PocketBase token
            scheme: bearer
            type: http
paths:
...
    /api/ext/users/{collection}/{id}/reset-password:
        post:
            description: Force-resets a user's password (no current password required). Invalidates all existing tokens. Superuser only.
            operationId: post_api_ext_users_collection_id_reset-password
            parameters:
                - in: path
                  name: collection
                  required: true
                  schema:
                    type: string
                - in: path
                  name: id
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Admin reset user password
            tags:
                - Auth
...